+++ src/main/java/org/kdb/inside/brains/view/console/watch/actions/AddWatchAction.java
@@ -1,0 +1,17 @@
+package org.kdb.inside.brains.view.console.watch.actions;
+
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import org.jetbrains.annotations.NotNull;
+import org.kdb.inside.brains.view.console.watch.WatchesView;
+
+public class AddWatchAction extends AnWatchesTreeAction {
+    @Override
+    protected void perform(@NotNull AnActionEvent e, @NotNull WatchesView watchesView) {
+        watchesView.addVariable();
+    }
+
+    @Override
+    protected boolean isEnabled(@NotNull AnActionEvent e, @NotNull WatchesView watchesView) {
+        return true;
+    }
+}